+++ test/conformance/exp_command_buffer/update/kernel_handle_update.cpp
@@ -269,7 +269,8 @@ TEST_P(urCommandBufferKernelHandleUpdateTest, Success) {
     ASSERT_SUCCESS(urCommandBufferAppendKernelLaunchExp(
         updatable_cmd_buf_handle, SaxpyKernel->Kernel, SaxpyKernel->NDimensions,
         &(SaxpyKernel->GlobalOffset), &(SaxpyKernel->GlobalSize),
-        &(SaxpyKernel->LocalSize), KernelAlternatives.size(),
+        &(SaxpyKernel->LocalSize),
+        static_cast<uint32_t>(KernelAlternatives.size()),
         KernelAlternatives.data(), 0, nullptr, 0, nullptr, nullptr, nullptr,
         CommandHandle.ptr()));
     ASSERT_NE(CommandHandle, nullptr);
@@ -298,7 +299,8 @@ TEST_P(urCommandBufferKernelHandleUpdateTest, UpdateAgain) {
     ASSERT_SUCCESS(urCommandBufferAppendKernelLaunchExp(
         updatable_cmd_buf_handle, SaxpyKernel->Kernel, SaxpyKernel->NDimensions,
         &(SaxpyKernel->GlobalOffset), &(SaxpyKernel->GlobalSize),
-        &(SaxpyKernel->LocalSize), KernelAlternatives.size(),
+        &(SaxpyKernel->LocalSize),
+        static_cast<uint32_t>(KernelAlternatives.size()),
         KernelAlternatives.data(), 0, nullptr, 0, nullptr, nullptr, nullptr,
         CommandHandle.ptr()));
     ASSERT_NE(CommandHandle, nullptr);
@@ -336,7 +338,8 @@ TEST_P(urCommandBufferKernelHandleUpdateTest, RestoreOriginalKernel) {
     ASSERT_SUCCESS(urCommandBufferAppendKernelLaunchExp(
         updatable_cmd_buf_handle, SaxpyKernel->Kernel, SaxpyKernel->NDimensions,
         &(SaxpyKernel->GlobalOffset), &(SaxpyKernel->GlobalSize),
-        &(SaxpyKernel->LocalSize), KernelAlternatives.size(),
+        &(SaxpyKernel->LocalSize),
+        static_cast<uint32_t>(KernelAlternatives.size()),
         KernelAlternatives.data(), 0, nullptr, 0, nullptr, nullptr, nullptr,
         CommandHandle.ptr()));
     ASSERT_NE(CommandHandle, nullptr);
@@ -394,9 +397,9 @@ TEST_P(urCommandBufferKernelHandleUpdateTest,
                          updatable_cmd_buf_handle, SaxpyKernel->Kernel,
                          SaxpyKernel->NDimensions, &(SaxpyKernel->GlobalOffset),
                          &(SaxpyKernel->GlobalSize), &(SaxpyKernel->LocalSize),
-                         KernelAlternatives.size(), KernelAlternatives.data(),
-                         0, nullptr, 0, nullptr, nullptr, nullptr,
-                         &CommandHandle));
+                         static_cast<uint32_t>(KernelAlternatives.size()),
+                         KernelAlternatives.data(), 0, nullptr, 0, nullptr,
+                         nullptr, nullptr, &CommandHandle));
 }
 
 using urCommandBufferValidUpdateParametersTest =
@@ -451,7 +454,8 @@ TEST_P(urCommandBufferValidUpdateParametersTest, UpdateOnlyLocalWorkSize) {
     ASSERT_SUCCESS(urCommandBufferAppendKernelLaunchExp(
         updatable_cmd_buf_handle, SaxpyKernel->Kernel, SaxpyKernel->NDimensions,
         &(SaxpyKernel->GlobalOffset), &(SaxpyKernel->GlobalSize),
-        &(SaxpyKernel->LocalSize), KernelAlternatives.size(),
+        &(SaxpyKernel->LocalSize),
+        static_cast<uint32_t>(KernelAlternatives.size()),
         KernelAlternatives.data(), 0, nullptr, 0, nullptr, nullptr, nullptr,
         CommandHandle.ptr()));
     ASSERT_NE(CommandHandle, nullptr);
